Added wrappings for torchvision dependencies, moved new tests to different test files, other cleanup
